Island: Rename and move file storage classes
