+++ monkey/monkey_island/cc/repository/__init__.py
@@ -1,0 +1,2 @@
+from .file_storage.filesystem_storage import FilesystemStorage
+from .file_storage.i_file_repository import IFileRepository, FileRetrievalError
