+++ monkey/monkey_island/cc/resources/pba_file_download.py
@@ -2,8 +2,8 @@ import logging
 
 from flask import make_response, send_file
 
+from monkey_island.cc.repository import FileRetrievalError, IFileRepository
 from monkey_island.cc.resources.AbstractResource import AbstractResource
-from monkey_island.cc.services import FileRetrievalError, IFileStorageService
 
 logger = logging.getLogger(__file__)
 
@@ -14,7 +14,7 @@ class PBAFileDownload(AbstractResource):
     File download endpoint used by monkey to download user's PBA file
     """
 
-    def __init__(self, file_storage_service: IFileStorageService):
+    def __init__(self, file_storage_service: IFileRepository):
         self._file_storage_service = file_storage_service
 
     # Used by monkey. can't secure.
