+++ monkey/monkey_island/cc/services/__init__.py
@@ -1,6 +1,3 @@
-from .i_file_storage_service import IFileStorageService, FileRetrievalError
-from .directory_file_storage_service import DirectoryFileStorageService
-
 from .authentication.authentication_service import AuthenticationService
 from .authentication.json_file_user_datastore import JsonFileUserDatastore
