+++ monkey/tests/unit_tests/monkey_island/cc/services/test_post_breach_files.py
@@ -4,13 +4,13 @@ import os
 import pytest
 from tests.utils import raise_
 
-from monkey_island.cc.services import DirectoryFileStorageService
+from monkey_island.cc.repository import FilesystemStorage
 from monkey_island.cc.services.post_breach_files import PostBreachFilesService
 
 
 @pytest.fixture
 def file_storage_service(tmp_path):
-    return DirectoryFileStorageService(tmp_path)
+    return FilesystemStorage(tmp_path)
 
 
 @pytest.fixture(autouse=True)
